feat: invalidate cloudfront caches when lambda configuration changes (#63)
*Issue #, if available:*

*Description of changes:*

When next.js contains a static page, changes to the page will not be
reflected due to cloudfront cache. This PR adds invalidation of old
cache when Lambda configuration changes


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.